+++ far/global.hpp
@@ -44,7 +44,7 @@ public:
 
 	HANDLE MainThreadHandle() const {return m_MainThreadHandle.native_handle();}
 	inline bool IsMainThread() const {return GetCurrentThreadId() == m_MainThreadId;}
-	uint64_t FarUpTime() const;
+	unsigned long long FarUpTime() const;
 	static const wchar_t* Version();
 	static const wchar_t* Copyright();
