+++ far/qview.hpp
@@ -52,7 +52,7 @@ public:
 private:
 	virtual int ProcessKey(const Manager::Key& Key) override;
 	virtual int ProcessMouse(const MOUSE_EVENT_RECORD *MouseEvent) override;
-	virtual __int64 VMProcess(int OpCode, void *vParam = nullptr, __int64 iParam = 0) override;
+	virtual long long VMProcess(int OpCode, void* vParam = nullptr, long long iParam = 0) override;
 	virtual void Update(int Mode) override;
 	virtual void CloseFile() override;
 	virtual void QViewDelTempName() override;
